Extract `priceFeeAmountForAsset`
into fees/utils
